==> notebooks/3.0-ak-bulding-predictive-model.ipynb <==
## Building Predictive Models

import pandas as pd
import os
import numpy as np

processed_data_path = os.path.join(os.path.pardir,'data','processed')
train_file_path = os.path.join(processed_data_path,'train.csv')
test_file_path = os.path.join(processed_data_path,'test.csv')

train_df = pd.read_csv(train_file_path, index_col='PassengerId')
test_df = pd.read_csv(test_file_path, index_col='PassengerId')

train_df.info()

test_df.info()

X = train_df.loc[:,'Age':].as_matrix().astype('float')
Y = train_df['Survived'].ravel()

X.shape, Y.shape

from sklearn.model_selection import train_test_split
X_train, X_test, Y_train, Y_test = train_test_split(X, Y, test_size=0.2, random_state=0)
print(X_train.shape, Y_train.shape)
print(X_test.shape, Y_test.shape)

print('mean survival in train: {0:.3f}'.format(np.mean(Y_train)))
print('mean survival in train: {0:.3f}'.format(np.mean(Y_test)))

### Check Scikit-Learn Version

import sklearn

sklearn.__version__

### Baseline Model

from sklearn.dummy import DummyClassifier

model_dummy = DummyClassifier(strategy='most_frequent', random_state=0)

model_dummy.fit(X_train, Y_train)

print('Score for Basline Model: {0:.2f}'.format(model_dummy.score(X_test,Y_test)))

from sklearn.metrics import accuracy_score, confusion_matrix, precision_score, recall_score

print('accuracy for baseline model: {0:.2f}'.format(accuracy_score(Y_test,model_dummy.predict(X_test))))

print('confusion matrix for basline model: \n {0}'.format(confusion_matrix(Y_test,model_dummy.predict(X_test))))

print('precision for baseline model: {0:.2f}'.format(precision_score(Y_test,model_dummy.predict(X_test))))
print('recall for baseline model: {0:.2f}'.format(recall_score(Y_test,model_dummy.predict(X_test))))

### First Kaggle Submission

test_X = test_df.as_matrix().astype('float')

predictions = model_dummy.predict(test_X)

df_submission = pd.DataFrame({'PassengerId': test_df.index, 'Survived': predictions})

df_submission.head()

submission_data_path = os.path.join(os.path.pardir, 'data', 'external')
submission_file_path = os.path.join(submission_data_path, '01_dummy.csv')

df_submission.to_csv(submission_file_path,index=False)

def get_submission_file(model, filename):
    test_X = test_df.as_matrix().astype('float')
    predictions = model.predict(test_X)
    df_submission = pd.DataFrame({'PassengerId': test_df.index, 'Survived': predictions})
    submission_data_path = os.path.join(os.path.pardir, 'data', 'external')
    submission_file_path = os.path.join(submission_data_path, filename)
    df_submission.to_csv(submission_file_path,index=False)

get_submission_file(model_dummy, '01_dummy.csv')

model_lr_1 = LogisticRegression(random_state=0)

model_lr_1.fit(X_train, Y_train)

print('Score for logistic regression - version 1 : {0:.2f}'.format(model_lr_1.score(X_test, Y_test)))

print('accuracy for baseline model: {0:.2f}'.format(accuracy_score(Y_test,model_lr_1.predict(X_test))))
print('confusion matrix for basline model: \n {0}'.format(confusion_matrix(Y_test,model_lr_1.predict(X_test))))
print('precision for baseline model: {0:.2f}'.format(precision_score(Y_test,model_lr_1.predict(X_test))))
print('recall for baseline model: {0:.2f}'.format(recall_score(Y_test,model_lr_1.predict(X_test))))

model_lr_1.coef_

### Second Kaggle Submission

get_submission_file(model_lr_1, '02_lr.csv')

## Part 2

### Hyperparameter Optimization

model_lr = LogisticRegression(random_state=0)

from sklearn.model_selection import GridSearchCV

parameters = {'C':[1.0,10.0,50.0,100.0,1000.0], 'penalty':['l1','l2']}
clf = GridSearchCV(model_lr, param_grid=parameters, cv=3)

clf.fit(X_train, Y_train)

clf.best_params_

print('best score : {0:.2f}'.format(clf.best_score_))

print('Score for logistic regression version 2: {0:.2f}'.format(clf.score(X_test, Y_test)))

### making third submission

get_submission_file(clf,'03_lr.csv')

### Feature Normalization and Standardization

### Feature Normalization

from sklearn.preprocessing import MinMaxScaler, StandardScaler

scaler = MinMaxScaler()
X_train_scaled = scaler.fit_transform(X_train)

X_train_scaled[:,0].min(), X_train_scaled[:,0].max()

X_test_scaled = scaler.transform(X_test)

### Feature Standardization

scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

### Create Model After Standardization

model_lr = LogisticRegression(random_state=0)
parameters = {'C':[1.0,10.0,50.0,100.0,1000.0], 'penalty':['l1','l2']}
clf = GridSearchCV(model_lr, param_grid=parameters, cv=3)
clf.fit(X_train, Y_train)

clf.best_score_

print('Score for logistic regression version 2: {0:.2f}'.format(clf.score(X_test, Y_test)))

### Model Persistence

import pickle

model_file_path = os.path.join(os.path.pardir, 'models', 'lr_model.pkl')
scaler_file_path = os.path.join(os.path.pardir, 'models', 'lr_scaler.pkl')

model_file_pickle = open(model_file_path, 'wb')
scaler_file_pickle = open(scaler_file_path, 'wb')

pickle.dump(clf, model_file_pickle)
pickle.dump(scaler, scaler_file_pickle)

model_file_pickle.close()
scaler_file_pickle.close()

#### load persisted files

model_file_pickle = open(model_file_path, 'rb')
scaler_file_pickle = open(scaler_file_path, 'rb')

clf_loaded = pickle.load(model_file_pickle)
scaler_loaded = pickle.load(scaler_file_pickle)

model_file_pickle.close()
scaler_file_pickle.close()

clf_loaded

scaler_loaded

X_test_scaled = scaler_loaded.transform(X_test)
print('Score for persisted logistic regression version 2: {0:.2f}'.format(clf_loaded.score(X_test, Y_test)))
